Working on cleaning the PR
